+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Feat: Collect more information for exceptions collected via `FlutterError.onError` (#538)
+* Feat: HTTP breadcrumbs have the request & response size if available (#552)
 
 # 6.0.0-beta.3